# lite_agent_web/lib/lite_agent_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "web",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/user/info": {
      "get": {
        "summary": "获取当前用户的信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "UserController"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataAccount",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "id": "",
                        "name": "",
                        "email": "",
                        "avatar": "",
                        "status": 0,
                        "createTime": "",
                        "updateTime": ""
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/user/update": {
      "put": {
        "summary": "修改用户信息",
        "deprecated": false,
        "description": "",
        "tags": [
          "UserController"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "昵称",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "avatar",
            "in": "query",
            "description": "头像url",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/user/updatePwd": {
      "put": {
        "summary": "修改密码",
        "deprecated": false,
        "description": "",
        "tags": [
          "UserController"
        ],
        "parameters": [
          {
            "name": "originPwd",
            "in": "query",
            "description": "旧密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newPwd",
            "in": "query",
            "description": "新密码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/user/resetPwd/captcha": {
      "post": {
        "summary": "重置密码-验证邮箱",
        "deprecated": false,
        "description": "",
        "tags": [
          "UserController"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "邮箱",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/user/resetPwd/captcha/verify": {
      "post": {
        "summary": "重置密码-检查验证码",
        "deprecated": false,
        "description": "",
        "tags": [
          "UserController"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "captcha",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/user/resetPwd/confirm": {
      "post": {
        "summary": "重置密码",
        "deprecated": false,
        "description": "",
        "tags": [
          "UserController"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "ResponseDataAccount": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "$ref": "#/components/schemas/Account",
            "description": "响应数据对象"
          }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": ""
          },
          "name": {
            "type": "string",
            "description": "昵称"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "status": {
            "type": "integer",
            "description": "账号状态"
          },
          "createTime": {
            "type": "string",
            "description": "",
            "x-apifox-mock": "@datetime"
          },
          "updateTime": {
            "type": "string",
            "description": "",
            "x-apifox-mock": "@datetime"
          },
          "deleted": {
            "type": "string",
            "description": "逻辑删除",
            "default": "0"
          }
        }
      },
...
      "ResponseDataString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "type": "string",
            "description": "响应数据对象"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": []
}